# tmp/flows/run-20260121T201110Z/bundles/finance-demo/openapi.yaml
components:
...
  securitySchemes:
    ActorHeader:
      in: header
      name: X-Actor-Id
      type: apiKey
info:
  title: Finance Pilot API
  version: idl.v1.2.2
openapi: 3.0.3
paths:
  /approvals/{id}/decide:
    post:
      operationId: decide_approval
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                decision:
                  enum:
                  - approve
                  - reject
                  type: string
              required:
              - decision
              type: object
        required: true
      responses:
        '200':
          description: Decision recorded
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: SoD violation
      summary: Decide on approval
      tags:
      - approvals
...